add docker scripts
