+++ docker/compose/movement-full-node/docker-compose.fullnode.yml
@@ -1,0 +1,22 @@
+version: "3.8"
+services:
+  movement-full-node:
+    image: ghcr.io/movementlabsxyz/movement-full-node:${CONTAINER_REV}
+    container_name: movement-full-node
+    command: run
+    environment:
+      - DOT_MOVEMENT_PATH=/.movement
+      - MOVEMENT_TIMING=info
+      - SUZUKA_TIMING_LOG=/.movement/movement-timing.log
+      - RUST_BACKTRACE=1
+    volumes:
+      - ${DOT_MOVEMENT_PATH}:/.movement
+    ports:
+      - "30731:30731"
+      - "30734:30734"
+    healthcheck:
+      test: [ "CMD-SHELL", "echo true" ]
+      retries: 10
+      interval: 10s
+      timeout: 5s
+    restart: on-failure:5
