+++ networks/movement/movement-full-node/src/da/mod.rs
@@ -1,3 +1,4 @@
+mod run;
 pub mod stream_blocks;
 
 use clap::Subcommand;
@@ -6,12 +7,14 @@ use clap::Subcommand;
 #[clap(rename_all = "kebab-case", about = "Commands for intereacting with the DA")]
 pub enum Da {
 	StreamBlocks(stream_blocks::StreamBlocks),
+	Run(run::DaRun),
 }
 
 impl Da {
 	pub async fn execute(&self) -> Result<(), anyhow::Error> {
 		match self {
 			Da::StreamBlocks(stream_blocks) => stream_blocks.execute().await,
+			Da::Run(da) => da.execute().await,
 		}
 	}
 }
