+++ networks/movement/movement-full-node/src/da/run.rs
@@ -1,0 +1,17 @@
+use crate::common_args::MovementArgs;
+use clap::Parser;
+
+#[derive(Debug, Parser, Clone)]
+#[clap(rename_all = "kebab-case", about = "Runs the Suzuka Full Node")]
+pub struct DaRun {
+	#[clap(flatten)]
+	pub movement_args: MovementArgs,
+}
+
+impl DaRun {
+	pub async fn execute(&self) -> Result<(), anyhow::Error> {
+		// get the config file
+		let dot_movement = self.movement_args.dot_movement()?;
+		movement_da_sequencer_node::start(dot_movement).await
+	}
+}
