+++ process-compose/movement-full-node/process-compose.yml
@@ -38,7 +38,7 @@ processes:
 
   da-sequencer:
     command: |
-      cargo run --bin movement-da-sequencer-node
+      RUST_BACKTRACE=1 movement-full-node da run
     depends_on:
       celestia-light-node:
         condition: process_healthy
